pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-glucorl"
version = "0.1.0"
description = "GlucoRL — Insulin Dosing RL Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.1",
    "simglucose>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools]
packages = ["glucorl", "glucorl.server"]
package-dir = {"glucorl" = ".", "glucorl.server" = "server"}

[project.scripts]
server = "server.app:main"
